Move advanced graph algorithm demos to the demos module (#866)
