+++ src/advanced_graph_algorithms/advanced_graph_algorithms.h
@@ -9,6 +9,8 @@ Graph* transpose_graph(Graph* graph);
 int** find_scc_tarjan(Graph* graph, int* scc_count, int** scc_sizes);
 int** find_scc_kosaraju(Graph* graph, int* scc_count, int** scc_sizes);
 void free_scc_result(int** sccs, int* scc_sizes, int scc_count);
+void visualize_tarjan(Graph* graph);
+void visualize_kosaraju(Graph* graph);
 void scc_demo(void);
 void tarjan_scc_demo(void);
 void kosaraju_scc_demo(void);
@@ -17,6 +19,9 @@ void kosaraju_scc_demo(void);
 int ford_fulkerson(weightedGraph* graph, int source, int sink);
 int edmonds_karp(weightedGraph* graph, int source, int sink);
 int dinic(weightedGraph* graph, int source, int sink);
+void visualize_ford_fulkerson(weightedGraph* graph, int source, int sink);
+void visualize_edmonds_karp(weightedGraph* graph, int source, int sink);
+void visualize_dinic(weightedGraph* graph, int source, int sink);
 void max_flow_demo(void);
 void ford_fulkerson_demo(void);
 void edmonds_karp_demo(void);
